Finished support for fuckitup extensions
