+++ ella/newman/widget_extensions.py
@@ -1,0 +1,82 @@
+'''
+Created on 3.11.2011
+
+@author: xaralis
+'''
+from ella.newman.conf import newman_settings
+
+__all__ = ['rich_text_extensions',]
+
+JS_EXTENSIONS = []
+CSS_EXTENSIONS = []
+
+class NewmanRichTextAreaExtensions(object):
+    """
+    Singleton object to keep track of RichTextArea extensions for markdown.
+    """
+    
+    def register(self, path, type='js'):
+        """
+        Add JS extension path to load for RichTextArea
+        """
+        if not type in ('js', 'css'):
+            raise ValueError('Only js or css types are supported.')
+        if type == 'js':
+            JS_EXTENSIONS.append(path)
+        else:
+            CSS_EXTENSIONS.append(path)
+            
+        print "register %s " % path
+    
+    def unregister(self, path, type):
+        """
+        Remove previously added JS extension path to load for RichTextArea
+        """
+        if not type in ('js', 'css'):
+            raise ValueError('Only js or css types are supported.')
+        if type == 'js':
+            if path in self.js_extensions:
+                JS_EXTENSIONS.remove(path)
+        else:
+            if path in self.css_extensions:
+                CSS_EXTENSIONS.remove(path)
+            
+    def get_js_extensions(self):
+        """
+        Get list of all JS extension paths
+        """
+        return JS_EXTENSIONS
+    
+    def get_css_extensions(self):
+        """
+        Get list of all CSS extension paths
+        """
+        return CSS_EXTENSIONS
+
+
+rich_text_extensions = NewmanRichTextAreaExtensions()
+
+try:
+    import markdown
+    rich_text_extensions.register(
+        newman_settings.MEDIA_PREFIX + 'js/fuckitup_extensions/markdown_tables.js',
+        type='js'
+    )
+    rich_text_extensions.register(
+        newman_settings.MEDIA_PREFIX + 'css/fuckitup_extensions/markdown_tables.css',
+        type='css'
+    )
+except ImportError:
+    try:
+        import markdown2
+        rich_text_extensions.register(
+            newman_settings.MEDIA_PREFIX + 'js/fuckitup_extensions/markdown2_tables.js',
+            type='js'
+        )
+        rich_text_extensions.register(
+            newman_settings.MEDIA_PREFIX + 'css/fuckitup_extensions/markdown_tables.css',
+            type='css'
+        )
+    except ImportError:
+        pass
+
